requirements.txt
Django>=5.0
psycopg2-binary>=2.9
celery>=5.4
redis>=5.0
requests>=2.32
djangorestframework>=3.15
python-dotenv>=1.0
gunicorn>=21.2